backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.129.0",
    "google-cloud-tasks>=2.21.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.12.0",
    "supabase>=2.28.0",
]
